[YDB-50]: Breaks up the large files.
