+++ YapDatabaseExtensions/Shared/Write.swift
@@ -110,581 +110,3 @@ extension SequenceType where Generator.Element: Persistable {
     }
 }
 
-// MARK: - Objects with no Metadata
-
-extension Writable
-    where
-    ItemType: NSCoding,
-    ItemType: Persistable {
-
-    /**
-    Write the items using an existing transaction.
-    
-    - parameter transaction: a YapDatabaseReadWriteTransaction
-    */
-    public func on(transaction: Database.Connection.WriteTransaction) {
-        items.forEach { transaction.writeAtIndex($0.index, object: $0, metadata: .None) }
-    }
-
-    /**
-    Write the items synchronously using a connection.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func sync(connection: Database.Connection) {
-        connection.write(on)
-    }
-
-    /**
-    Write the items asynchronously using a connection.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func async(connection: Database.Connection, queue: dispatch_queue_t = dispatch_get_main_queue(), completion: dispatch_block_t) {
-        connection.asyncWrite(on, queue: queue, completion: completion)
-    }
-
-    /**
-    Write the items inside of an `NSOperation`.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func operation(connection: Database.Connection) -> NSOperation {
-        return connection.writeBlockOperation { self.on($0) }
-    }
-}
-
-// MARK: - Objects with Object Metadata
-
-extension Writable
-    where
-    ItemType: NSCoding,
-    ItemType: MetadataPersistable,
-    ItemType.MetadataType: NSCoding {
-
-    /**
-    Write the items using an existing transaction.
-
-    - parameter transaction: a YapDatabaseReadWriteTransaction
-    */
-    public func on(transaction: Database.Connection.WriteTransaction) {
-        transaction.write(items)
-    }
-
-    /**
-    Write the items synchronously using a connection.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func sync(connection: Database.Connection) {
-        connection.write(on)
-    }
-
-    /**
-    Write the items asynchronously using a connection.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func async(connection: Database.Connection, queue: dispatch_queue_t = dispatch_get_main_queue(), completion: dispatch_block_t) {
-        connection.asyncWrite(on, queue: queue, completion: completion)
-    }
-
-    /**
-    Write the items inside of an `NSOperation`.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func operation(connection: Database.Connection) -> NSOperation {
-        return connection.writeBlockOperation { self.on($0) }
-    }
-}
-
-// MARK: - Objects with Value Metadata
-
-extension Writable
-    where
-    ItemType: NSCoding,
-    ItemType: MetadataPersistable,
-    ItemType.MetadataType: ValueCoding,
-    ItemType.MetadataType.Coder: NSCoding,
-    ItemType.MetadataType.Coder.ValueType == ItemType.MetadataType {
-
-    /**
-    Write the items using an existing transaction.
-
-    - parameter transaction: a YapDatabaseReadWriteTransaction
-    */
-    public func on(transaction: Database.Connection.WriteTransaction) {
-        transaction.write(items)
-    }
-
-    /**
-    Write the items synchronously using a connection.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func sync(connection: Database.Connection) {
-        connection.write(on)
-    }
-
-    /**
-    Write the items asynchronously using a connection.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func async(connection: Database.Connection, queue: dispatch_queue_t = dispatch_get_main_queue(), completion: dispatch_block_t) {
-        connection.asyncWrite(on, queue: queue, completion: completion)
-    }
-
-    /**
-    Write the items inside of an `NSOperation`.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func operation(connection: Database.Connection) -> NSOperation {
-        return connection.writeBlockOperation { self.on($0) }
-    }
-}
-
-// MARK: - Values with no Metadata
-
-extension Writable
-    where
-    ItemType: ValueCoding,
-    ItemType: Persistable,
-    ItemType.Coder: NSCoding,
-    ItemType.Coder.ValueType == ItemType {
-
-    /**
-    Write the items using an existing transaction.
-
-    - parameter transaction: a YapDatabaseReadWriteTransaction
-    */
-    public func on(transaction: Database.Connection.WriteTransaction) {
-        items.forEach { transaction.writeAtIndex($0.index, object: $0.encoded, metadata: .None) }
-    }
-
-    /**
-    Write the items synchronously using a connection.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func sync(connection: Database.Connection) {
-        connection.write(on)
-    }
-
-    /**
-    Write the items asynchronously using a connection.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func async(connection: Database.Connection, queue: dispatch_queue_t = dispatch_get_main_queue(), completion: dispatch_block_t) {
-        connection.asyncWrite(on, queue: queue, completion: completion)
-    }
-
-    /**
-    Write the items inside of an `NSOperation`.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func operation(connection: Database.Connection) -> NSOperation {
-        return connection.writeBlockOperation { self.on($0) }
-    }
-}
-
-// MARK: - Values with Object Metadata
-
-extension Writable
-    where
-    ItemType: ValueCoding,
-    ItemType: MetadataPersistable,
-    ItemType.Coder: NSCoding,
-    ItemType.Coder.ValueType == ItemType,
-    ItemType.MetadataType: NSCoding {
-
-    /**
-    Write the items using an existing transaction.
-
-    - parameter transaction: a YapDatabaseReadWriteTransaction
-    */
-    public func on(transaction: Database.Connection.WriteTransaction) {
-        transaction.write(items)
-    }
-
-    /**
-    Write the items synchronously using a connection.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func sync(connection: Database.Connection) {
-        connection.write(on)
-    }
-
-    /**
-    Write the items asynchronously using a connection.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func async(connection: Database.Connection, queue: dispatch_queue_t = dispatch_get_main_queue(), completion: dispatch_block_t) {
-        connection.asyncWrite(on, queue: queue, completion: completion)
-    }
-
-    /**
-    Write the items inside of an `NSOperation`.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func operation(connection: Database.Connection) -> NSOperation {
-        return connection.writeBlockOperation { self.on($0) }
-    }
-}
-
-// MARK: - Values with Value Metadata
-
-extension Writable
-    where
-    ItemType: ValueCoding,
-    ItemType: MetadataPersistable,
-    ItemType.Coder: NSCoding,
-    ItemType.Coder.ValueType == ItemType,
-    ItemType.MetadataType: ValueCoding,
-    ItemType.MetadataType.Coder: NSCoding,
-    ItemType.MetadataType.Coder.ValueType == ItemType.MetadataType {
-
-    /**
-    Write the items using an existing transaction.
-
-    - parameter transaction: a YapDatabaseReadWriteTransaction
-    */
-    public func on(transaction: Database.Connection.WriteTransaction) {
-        transaction.write(items)
-    }
-
-    /**
-    Write the items synchronously using a connection.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func sync(connection: Database.Connection) {
-        connection.write(on)
-    }
-
-    /**
-    Write the items asynchronously using a connection.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func async(connection: Database.Connection, queue: dispatch_queue_t = dispatch_get_main_queue(), completion: dispatch_block_t) {
-        connection.asyncWrite(on, queue: queue, completion: completion)
-    }
-
-    /**
-    Write the items inside of an `NSOperation`.
-
-    - parameter connection: a YapDatabaseConnection
-    */
-    public func operation(connection: Database.Connection) -> NSOperation {
-        return connection.writeBlockOperation { self.on($0) }
-    }
-}
-
-
-// MARK: - Object with Object metadata
-
-extension WriteTransactionType {
-
-    public func write<
-        ObjectWithObjectMetadata
-        where
-        ObjectWithObjectMetadata: MetadataPersistable,
-        ObjectWithObjectMetadata: NSCoding,
-        ObjectWithObjectMetadata.MetadataType: NSCoding>(item: ObjectWithObjectMetadata) {
-            writeAtIndex(item.index, object: item, metadata: item.metadata)
-    }
-
-    public func write<
-        ObjectWithObjectMetadata
-        where
-        ObjectWithObjectMetadata: MetadataPersistable,
-        ObjectWithObjectMetadata: NSCoding,
-        ObjectWithObjectMetadata.MetadataType: NSCoding>(items: [ObjectWithObjectMetadata]) {
-            items.forEach(write)
-    }
-}
-
-extension ConnectionType {
-
-    public func write<
-        ObjectWithObjectMetadata
-        where
-        ObjectWithObjectMetadata: MetadataPersistable,
-        ObjectWithObjectMetadata: NSCoding,
-        ObjectWithObjectMetadata.MetadataType: NSCoding>(item: ObjectWithObjectMetadata) {
-            write { $0.write(item) }
-    }
-
-    public func write<
-        ObjectWithObjectMetadata
-        where
-        ObjectWithObjectMetadata: MetadataPersistable,
-        ObjectWithObjectMetadata: NSCoding,
-        ObjectWithObjectMetadata.MetadataType: NSCoding>(items: [ObjectWithObjectMetadata]) {
-            write { $0.write(items) }
-    }
-
-    public func asyncWrite<
-        ObjectWithObjectMetadata
-        where
-        ObjectWithObjectMetadata: MetadataPersistable,
-        ObjectWithObjectMetadata: NSCoding,
-        ObjectWithObjectMetadata.MetadataType: NSCoding>(item: ObjectWithObjectMetadata, queue: dispatch_queue_t = dispatch_get_main_queue(), completion: dispatch_block_t) {
-            asyncWrite({ $0.write(item) }, queue: queue, completion: { _ in completion() })
-    }
-
-    public func asyncWrite<
-        ObjectWithObjectMetadata
-        where
-        ObjectWithObjectMetadata: MetadataPersistable,
-        ObjectWithObjectMetadata: NSCoding,
-        ObjectWithObjectMetadata.MetadataType: NSCoding>(items: [ObjectWithObjectMetadata], queue: dispatch_queue_t = dispatch_get_main_queue(), completion: dispatch_block_t) {
-            asyncWrite({ $0.write(items) }, queue: queue, completion: { _ in completion() })
-    }
-}
-
-// MARK: - Object with Value metadata
-
-extension WriteTransactionType {
-
-    public func write<
-        ObjectWithValueMetadata
-        where
-        ObjectWithValueMetadata: MetadataPersistable,
-        ObjectWithValueMetadata: NSCoding,
-        ObjectWithValueMetadata.MetadataType: ValueCoding,
-        ObjectWithValueMetadata.MetadataType.Coder: NSCoding,
-        ObjectWithValueMetadata.MetadataType.Coder.ValueType == ObjectWithValueMetadata.MetadataType>(item: ObjectWithValueMetadata) {
-            writeAtIndex(item.index, object: item, metadata: item.metadata?.encoded)
-    }
-
-    public func write<
-        ObjectWithValueMetadata
-        where
-        ObjectWithValueMetadata: MetadataPersistable,
-        ObjectWithValueMetadata: NSCoding,
-        ObjectWithValueMetadata.MetadataType: ValueCoding,
-        ObjectWithValueMetadata.MetadataType.Coder: NSCoding,
-        ObjectWithValueMetadata.MetadataType.Coder.ValueType == ObjectWithValueMetadata.MetadataType>(items: [ObjectWithValueMetadata]) {
-            items.forEach(write)
-    }
-}
-
-extension ConnectionType {
-
-    public func write<
-        ObjectWithValueMetadata
-        where
-        ObjectWithValueMetadata: MetadataPersistable,
-        ObjectWithValueMetadata: NSCoding,
-        ObjectWithValueMetadata.MetadataType: ValueCoding,
-        ObjectWithValueMetadata.MetadataType.Coder: NSCoding,
-        ObjectWithValueMetadata.MetadataType.Coder.ValueType == ObjectWithValueMetadata.MetadataType>(item: ObjectWithValueMetadata) {
-            write { $0.write(item) }
-    }
-
-    public func write<
-        ObjectWithValueMetadata
-        where
-        ObjectWithValueMetadata: MetadataPersistable,
-        ObjectWithValueMetadata: NSCoding,
-        ObjectWithValueMetadata.MetadataType: ValueCoding,
-        ObjectWithValueMetadata.MetadataType.Coder: NSCoding,
-        ObjectWithValueMetadata.MetadataType.Coder.ValueType == ObjectWithValueMetadata.MetadataType>(items: [ObjectWithValueMetadata]) {
-            write { $0.write(items) }
-    }
-
-    public func asyncWrite<
-        ObjectWithValueMetadata
-        where
-        ObjectWithValueMetadata: MetadataPersistable,
-        ObjectWithValueMetadata: NSCoding,
-        ObjectWithValueMetadata.MetadataType: ValueCoding,
-        ObjectWithValueMetadata.MetadataType.Coder: NSCoding,
-        ObjectWithValueMetadata.MetadataType.Coder.ValueType == ObjectWithValueMetadata.MetadataType>(item: ObjectWithValueMetadata, queue: dispatch_queue_t = dispatch_get_main_queue(), completion: dispatch_block_t) {
-            asyncWrite({ $0.write(item) }, queue: queue, completion: { _ in completion() })
-    }
-
-    public func asyncWrite<
-        ObjectWithValueMetadata
-        where
-        ObjectWithValueMetadata: MetadataPersistable,
-        ObjectWithValueMetadata: NSCoding,
-        ObjectWithValueMetadata.MetadataType: ValueCoding,
-        ObjectWithValueMetadata.MetadataType.Coder: NSCoding,
-        ObjectWithValueMetadata.MetadataType.Coder.ValueType == ObjectWithValueMetadata.MetadataType>(items: [ObjectWithValueMetadata], queue: dispatch_queue_t = dispatch_get_main_queue(), completion: dispatch_block_t) {
-            asyncWrite({ $0.write(items) }, queue: queue, completion: { _ in completion() })
-    }
-}
-
-// MARK: - Value with Object metadata
-
-extension WriteTransactionType {
-
-    public func write<
-        ValueWithObjectMetadata
-        where
-        ValueWithObjectMetadata: MetadataPersistable,
-        ValueWithObjectMetadata: ValueCoding,
-        ValueWithObjectMetadata.Coder: NSCoding,
-        ValueWithObjectMetadata.Coder.ValueType == ValueWithObjectMetadata,
-        ValueWithObjectMetadata.MetadataType: NSCoding>(item: ValueWithObjectMetadata) {
-            writeAtIndex(item.index, object: item.encoded, metadata: item.metadata)
-    }
-
-    public func write<
-        ValueWithObjectMetadata
-        where
-        ValueWithObjectMetadata: MetadataPersistable,
-        ValueWithObjectMetadata: ValueCoding,
-        ValueWithObjectMetadata.Coder: NSCoding,
-        ValueWithObjectMetadata.Coder.ValueType == ValueWithObjectMetadata,
-        ValueWithObjectMetadata.MetadataType: NSCoding>(items: [ValueWithObjectMetadata]) {
-            items.forEach(write)
-    }
-}
-
-extension ConnectionType {
-
-    public func write<
-        ValueWithObjectMetadata
-        where
-        ValueWithObjectMetadata: MetadataPersistable,
-        ValueWithObjectMetadata: ValueCoding,
-        ValueWithObjectMetadata.Coder: NSCoding,
-        ValueWithObjectMetadata.Coder.ValueType == ValueWithObjectMetadata,
-        ValueWithObjectMetadata.MetadataType: NSCoding>(item: ValueWithObjectMetadata) {
-            write { $0.write(item) }
-    }
-
-    public func write<
-        ValueWithObjectMetadata
-        where
-        ValueWithObjectMetadata: MetadataPersistable,
-        ValueWithObjectMetadata: ValueCoding,
-        ValueWithObjectMetadata.Coder: NSCoding,
-        ValueWithObjectMetadata.Coder.ValueType == ValueWithObjectMetadata,
-        ValueWithObjectMetadata.MetadataType: NSCoding>(items: [ValueWithObjectMetadata]) {
-            write { $0.write(items) }
-    }
-
-    public func asyncWrite<
-        ValueWithObjectMetadata
-        where
-        ValueWithObjectMetadata: MetadataPersistable,
-        ValueWithObjectMetadata: ValueCoding,
-        ValueWithObjectMetadata.Coder: NSCoding,
-        ValueWithObjectMetadata.Coder.ValueType == ValueWithObjectMetadata,
-        ValueWithObjectMetadata.MetadataType: NSCoding>(item: ValueWithObjectMetadata, queue: dispatch_queue_t = dispatch_get_main_queue(), completion: dispatch_block_t) {
-            asyncWrite({ $0.write(item) }, queue: queue, completion: { _ in completion() })
-    }
-
-    public func asyncWrite<
-        ValueWithObjectMetadata
-        where
-        ValueWithObjectMetadata: MetadataPersistable,
-        ValueWithObjectMetadata: ValueCoding,
-        ValueWithObjectMetadata.Coder: NSCoding,
-        ValueWithObjectMetadata.Coder.ValueType == ValueWithObjectMetadata,
-        ValueWithObjectMetadata.MetadataType: NSCoding>(items: [ValueWithObjectMetadata], queue: dispatch_queue_t = dispatch_get_main_queue(), completion: dispatch_block_t) {
-            asyncWrite({ $0.write(items) }, queue: queue, completion: { _ in completion() })
-    }
-}
-
-// MARK: - Value with Value metadata
-
-extension WriteTransactionType {
-
-    public func write<
-        ValueWithValueMetadata
-        where
-        ValueWithValueMetadata: MetadataPersistable,
-        ValueWithValueMetadata: ValueCoding,
-        ValueWithValueMetadata.Coder: NSCoding,
-        ValueWithValueMetadata.Coder.ValueType == ValueWithValueMetadata,
-        ValueWithValueMetadata.MetadataType: ValueCoding,
-        ValueWithValueMetadata.MetadataType.Coder: NSCoding,
-        ValueWithValueMetadata.MetadataType.Coder.ValueType == ValueWithValueMetadata.MetadataType>(item: ValueWithValueMetadata) {
-            writeAtIndex(item.index, object: item.encoded, metadata: item.metadata?.encoded)
-    }
-
-    public func write<
-        ValueWithValueMetadata
-        where
-        ValueWithValueMetadata: MetadataPersistable,
-        ValueWithValueMetadata: ValueCoding,
-        ValueWithValueMetadata.Coder: NSCoding,
-        ValueWithValueMetadata.Coder.ValueType == ValueWithValueMetadata,
-        ValueWithValueMetadata.MetadataType: ValueCoding,
-        ValueWithValueMetadata.MetadataType.Coder: NSCoding,
-        ValueWithValueMetadata.MetadataType.Coder.ValueType == ValueWithValueMetadata.MetadataType>(items: [ValueWithValueMetadata]) {
-            items.forEach(write)
-    }
-}
-
-extension ConnectionType {
-
-    public func write<
-        ValueWithValueMetadata
-        where
-        ValueWithValueMetadata: MetadataPersistable,
-        ValueWithValueMetadata: ValueCoding,
-        ValueWithValueMetadata.Coder: NSCoding,
-        ValueWithValueMetadata.Coder.ValueType == ValueWithValueMetadata,
-        ValueWithValueMetadata.MetadataType: ValueCoding,
-        ValueWithValueMetadata.MetadataType.Coder: NSCoding,
-        ValueWithValueMetadata.MetadataType.Coder.ValueType == ValueWithValueMetadata.MetadataType>(item: ValueWithValueMetadata) {
-            write { $0.write(item) }
-    }
-
-    public func write<
-        ValueWithValueMetadata
-        where
-        ValueWithValueMetadata: MetadataPersistable,
-        ValueWithValueMetadata: ValueCoding,
-        ValueWithValueMetadata.Coder: NSCoding,
-        ValueWithValueMetadata.Coder.ValueType == ValueWithValueMetadata,
-        ValueWithValueMetadata.MetadataType: ValueCoding,
-        ValueWithValueMetadata.MetadataType.Coder: NSCoding,
-        ValueWithValueMetadata.MetadataType.Coder.ValueType == ValueWithValueMetadata.MetadataType>(items: [ValueWithValueMetadata]) {
-            write { $0.write(items) }
-    }
-
-    public func asyncWrite<
-        ValueWithValueMetadata
-        where
-        ValueWithValueMetadata: MetadataPersistable,
-        ValueWithValueMetadata: ValueCoding,
-        ValueWithValueMetadata.Coder: NSCoding,
-        ValueWithValueMetadata.Coder.ValueType == ValueWithValueMetadata,
-        ValueWithValueMetadata.MetadataType: ValueCoding,
-        ValueWithValueMetadata.MetadataType.Coder: NSCoding,
-        ValueWithValueMetadata.MetadataType.Coder.ValueType == ValueWithValueMetadata.MetadataType>(item: ValueWithValueMetadata, queue: dispatch_queue_t = dispatch_get_main_queue(), completion: dispatch_block_t) {
-            asyncWrite({ $0.write(item) }, queue: queue, completion: { _ in completion() })
-    }
-
-    public func asyncWrite<
-        ValueWithValueMetadata
-        where
-        ValueWithValueMetadata: MetadataPersistable,
-        ValueWithValueMetadata: ValueCoding,
-        ValueWithValueMetadata.Coder: NSCoding,
-        ValueWithValueMetadata.Coder.ValueType == ValueWithValueMetadata,
-        ValueWithValueMetadata.MetadataType: ValueCoding,
-        ValueWithValueMetadata.MetadataType.Coder: NSCoding,
-        ValueWithValueMetadata.MetadataType.Coder.ValueType == ValueWithValueMetadata.MetadataType>(items: [ValueWithValueMetadata], queue: dispatch_queue_t = dispatch_get_main_queue(), completion: dispatch_block_t) {
-            asyncWrite({ $0.write(items) }, queue: queue, completion: { _ in completion() })
-    }
-}
-
-
-
-
