+++ YapDatabaseExtensions/Shared/Write_ValueWithNoMetadata.swift
@@ -1,0 +1,60 @@
+//
+//  Write_ValueWithNoMetadata.swift
+//  YapDatabaseExtensions
+//
+//  Created by Daniel Thorpe on 11/10/2015.
+//
+//
+
+import Foundation
+import ValueCoding
+import YapDatabase
+
+// MARK: - Values with no Metadata
+
+extension Writable
+    where
+    ItemType: ValueCoding,
+    ItemType: Persistable,
+    ItemType.Coder: NSCoding,
+    ItemType.Coder.ValueType == ItemType {
+
+    /**
+    Write the items using an existing transaction.
+
+    - parameter transaction: a YapDatabaseReadWriteTransaction
+    */
+    public func on(transaction: Database.Connection.WriteTransaction) {
+        items.forEach { transaction.writeAtIndex($0.index, object: $0.encoded, metadata: .None) }
+    }
+
+    /**
+    Write the items synchronously using a connection.
+
+    - parameter connection: a YapDatabaseConnection
+    */
+    public func sync(connection: Database.Connection) {
+        connection.write(on)
+    }
+
+    /**
+    Write the items asynchronously using a connection.
+
+    - parameter connection: a YapDatabaseConnection
+    */
+    public func async(connection: Database.Connection, queue: dispatch_queue_t = dispatch_get_main_queue(), completion: dispatch_block_t) {
+        connection.asyncWrite(on, queue: queue, completion: completion)
+    }
+
+    /**
+    Write the items inside of an `NSOperation`.
+
+    - parameter connection: a YapDatabaseConnection
+    */
+    public func operation(connection: Database.Connection) -> NSOperation {
+        return connection.writeBlockOperation { self.on($0) }
+    }
+}
+
+
+
